fix: Resolve Failing Login and Admin Studies Page Feedback (#664)
* fix: remove site_curator

* fix: resolve delete study record failing on not found

* feat: show dacName instead of dacId admin page

* fix: add data_admin and remove duplicated zod validation

* fix: change order of delete studies from import

* fix: remove AUTHZ_GROUP_ADMIN

* chore: update comment verbage
