+++ apps/api/.env.docker.schema
@@ -7,6 +7,5 @@ AUTH_CLIENT_SECRET=placeholder-please-replace
 
 # ## AUTHZ ###
 AUTHZ_ENDPOINT=https://auth.dev.pcgl.sd4h.ca/authz
-AUTHZ_GROUP_ADMIN=PCGL:data-admin
 AUTHZ_SERVICE_ID=service-id
 AUTHZ_SERVICE_UUID=00000000-0000-0000-0000-000000000000
