+++ apps/api/.env.schema
@@ -35,7 +35,6 @@ AUTHZ_FETCH_TIMEOUT_MS=10000
 APPROVED_PERMISSION_EXPIRES_IN_DAYS=365
 
 # -- Groups --
-AUTHZ_GROUP_ADMIN=PCGL:TEST-ADMIN
 AUTHZ_GROUP_PREFIX_DAC_CHAIR=PCGL:TEST:DAC:CHAIR:
 AUTHZ_GROUP_PREFIX_DAC_MEMBER=PCGL:TEST:DAC:MEMBER:
