+++ apps/api/README.md
@@ -39,7 +39,6 @@ The following sub-sections list the environment variables that can be set to con
 | `AUTHZ_FETCH_RETRIES`      | Number of times to retry a failed fetch request URL.                                                                                                                                                                                                                                                                                     | `number`  | Optional                                  | `2`        |
 | `AUTHZ_FETCH_RETRY_DELAY_MS` | Delay in milliseconds between retry attempts. URL.                                                                                                                                                                                                                                                                                        | `number`  | Optional                                  | `500`        |
 | `AUTHZ_FETCH_TIMEOUT_MS`   | Maximum time in milliseconds to wait for a fetch request before timing out. URL.                                                                                                                                                                                                                                                            | `number`  | Optional                                  | `10000`        |
-| `AUTHZ_GROUP_ADMIN`  | Authorization group for admins.                                                                                                                                                                                                                                                                                                            | `string`  | Conditional <br/> `DISABLE_AUTH !== true`                                  |            |
 | `AUTHZ_GROUP_PREFIX_DAC_MEMBER` | Authorization group for DAC members.                                                                                                                                                                                                                                                                                                   | `string`  | Conditional <br/> `DISABLE_AUTH !== true` |            |
 | `AUTHZ_GROUP_PREFIX_DAC_CHAIR` | Authorization group for DAC chair.                                                                                                                                                                                                                                                                                                      | `string`  | Conditional <br/> `DISABLE_AUTH !== true` |            |
 | `AUTHZ_SERVICE_ID`   | Authorization Service ID.                                                                                                                                                                                                                                                                                                                  | `string`  | Conditional <br/> `DISABLE_AUTH !== true` |            |
@@ -79,7 +78,6 @@ The following sub-sections list the environment variables that can be set to con
 | `AUTH_PROVIDER_HOST`        | The domain or IP address of the authentication server. Required when `AUTH_ENABLED` is enabled.                                                                                                                                                                                                                                                                                                             |                                        |
 | `AUTH_PROTECT_METHODS`      | Specifies a list of HTTP methods to protect, separated by commas. (Example: DELETE,GET,POST,PUT).                                                                                                                                                                                                                                                                                                           | 'DELETE,GET,POST,PUT'                  |
 | `AUTHZ_ENDPOINT`            | The domain or IP address of the authorization server. Required when `AUTH_ENABLED` is enabled.                                                                                                                                                                                                                                                                                                              |                                        |
-| `AUTHZ_GROUP_ADMIN`         | Authorization group configuration for determining admin group                                                                                                                                                                                                                                                                                                                                               |                                        |
 | `AUTHZ_SERVICE_ID`         | Authorization Service ID.                                                                                                                                                                                                                                                                                                                                               |                                        |
 | `AUTHZ_SERVICE_UUID`         | Authorization Service UUID.
