+++ apps/api/src/external/oidcAuthNClient.ts
@@ -28,7 +28,7 @@ import {
 	oidcTokenResponseSchema,
 	type OIDCUserInfoResponse,
 	oidcUserInfoResponseSchema,
-} from './types.ts';
+} from '@pcgl-daco/validation';
 
 const logger = BaseLogger.forModule('oidcAuthClient');
