+++ apps/api/tests/utils/.env.test
@@ -23,7 +23,6 @@ AUTH_CLIENT_SECRET=placeholder-please-replace
 
 # AUTHZ
 AUTHZ_ENDPOINT=https://auth.dev.pcgl.sd4h.ca/authz
-AUTHZ_GROUP_ADMIN=PCGL:data-admin
 AUTHZ_GROUP_PREFIX_DAC_CHAIR=PCGL:DACO:DAC-Chair:
 AUTHZ_GROUP_PREFIX_DAC_MEMBER=PCGL:DACO:DAC-Member:
 AUTHZ_SERVICE_ID=service-id
